+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Note that this project is currently in a beta state. Therefore, there might be API changes that are not reflected in the versioning.
 
+## [Unreleased]
+
+### Added
+
+- Added Cubemap Array textures
+- Possibility to create textures from device tensors
+
+### Changed
+
+- The Renderer now internally also uses the Texture ID push system
+- Cube maps are no longer interpreted as depth 6 textures
+
+### Fixed
+
+- Screenshot functionality of Renderer uses correct member method and not global renderer instance.
+
 ## [0.1.1-beta]
 
 ### Added